+++ src-tauri/src/utils/io/mod.rs
@@ -1,2 +1,3 @@
 pub mod file_helper;
 pub mod network;
+pub mod terminal;