+++ src/app/shared/overviews-shared/status-overview-panel/status-overview-panel.component.ts
@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
@@ -7,7 +6,7 @@ import { AppTab } from '../../components/types';
 @Component({
   selector: 'app-status-overview-panel',
   standalone: true,
-  imports: [CommonModule, MatCardModule, MatIconModule],
+  imports: [MatCardModule, MatIconModule],
   templateUrl: './status-overview-panel.component.html',
   styleUrl: './status-overview-panel.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,