+++ src/app/shared/remote-config/components/remote-config-step/remote-config-step.component.ts
@@ -9,14 +9,12 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
-import { CommonModule } from '@angular/common';
 
 import { SENSITIVE_KEYS } from '../../../components/types';
 
 @Component({
   selector: 'app-remote-config-step',
   imports: [
-    CommonModule,
     ReactiveFormsModule,
     MatFormFieldModule,
     MatSelectModule,