+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 # All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- Native console support for the native terminal. You can now open the remote configuration in the native terminal by clicking the "Remote Terminal" button in the top left add button. It will use the preferred terminal app from the settings. Also, you can set the preferred terminal app in the settings.
+
 ## [beta-0.1.2] - 2025-07-15
 ### Added
 - General tab added.